refactor: move controller and service
