+++ src/lib/routes/admin-api/project/project-archive.ts
@@ -9,7 +9,7 @@ import {
 import { Logger } from '../../../logger';
 import { extractUsername } from '../../../util/extract-user';
 import { DELETE_FEATURE } from '../../../types/permissions';
-import FeatureToggleService from '../../../services/feature-toggle-service';
+import FeatureToggleService from '../../../features/feature-toggle/feature-toggle-service';
 import { IAuthRequest } from '../../unleash-types';
 import { OpenApiService } from '../../../services/openapi-service';
 import {
