+++ src/lib/routes/admin-api/project/variants.ts
@@ -1,4 +1,4 @@
-import FeatureToggleService from '../../../services/feature-toggle-service';
+import FeatureToggleService from '../../../features/feature-toggle/feature-toggle-service';
 import { Logger } from '../../../logger';
 import Controller from '../../controller';
 import { IUnleashConfig } from '../../../types/option';
